Abstract subscribe link methods into concern
